Refactored the game worker package to be more structured like game.
