+++ aimmo-game-worker/simulation/avatar/avatar_state.py
@@ -1,4 +1,4 @@
-from simulation.location import Location
+from simulation.geography.location import Location
 
 
 class AvatarState(object):
