+++ aimmo-game-worker/tests/simulation/geography/test_location.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 from unittest import TestCase
 
-from simulation.location import Location
+from simulation.geography.location import Location
 
 
 class TestLocation(TestCase):
