+++ aimmo-game-worker/tests/simulation/test_action.py
@@ -1,8 +1,9 @@
 from __future__ import absolute_import
+
 from unittest import TestCase
 
 from simulation.action import WaitAction, MoveAction, AttackAction
-from simulation.direction import NORTH
+from simulation.geography.direction import NORTH
 
 
 class TestAction(TestCase):
